backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.10.6
redis==5.2.1
PyJWT==2.10.1
